app/requirements.txt
Flask
SQLAlchemy
psycopg[binary]
fastapi
uvicorn
fastapi
pika
jinja2
python-multipart
pytest
requests

torch
torchvision
pillow
python-multipart